+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Bug #61: Correct parameter order in docblocks for `BaseInput` and `BaseInline` classes (@terabytesoftw)
 - Bug #62: Reorder use statements in `BaseInput` class for better organization (@terabytesoftw)
 - Bug #63: Use `HasName` trait in `BaseInput` class to manage the `name` attribute (@terabytesoftw)
+- Enh #64: Add support for automatic `aria-describedby` attribute in `BaseInput` class (@terabytesoftw)
 
 ## 0.5.2 January 28, 2026